Update redistribution license.
This change clarifies the mixed-license nature of our source tree, and clarifies the restrictions
to official branding when built from source on target machines.
This change also excludes BSD and ReactOS from default exemption rules because the
respective target systems either do not follow standard Open Source licensing practices
as regards rebranding of modified sources, or have chosen to stay on a build target that
is no longer supported, respectively.